+++ src/khoj/interface/web/content_source_github_input.html
@@ -165,7 +165,7 @@
 
         // Save Github config on server
         const csrfToken = document.cookie.split('; ').find(row => row.startsWith('csrftoken'))?.split('=')[1];
-        fetch('/api/configure/content/github', {
+        fetch('/api/content/github', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
