+++ tests/test_client.py
@@ -269,11 +269,11 @@ def test_get_api_config_types(client, sample_org_data, default_user: KhojUser):
     text_search.setup(OrgToEntries, sample_org_data, regenerate=False, user=default_user)
 
     # Act
-    response = client.get(f"/api/configure/types", headers=headers)
+    response = client.get(f"/api/content/types", headers=headers)
 
     # Assert
     assert response.status_code == 200
-    assert response.json() == ["all", "org", "plaintext"]
+    assert set(response.json()) == {"all", "org", "plaintext"}
 
 
 # ----------------------------------------------------------------------------------------------------
@@ -289,7 +289,7 @@ def test_get_configured_types_with_no_content_config(fastapi_app: FastAPI):
     client = TestClient(fastapi_app)
 
     # Act
-    response = client.get(f"/api/configure/types")
+    response = client.get(f"/api/content/types")
 
     # Assert
     assert response.status_code == 200
